requirements.txt
Django==5.1.4
black==24.10.0
whitenoise[brotli]==6.6.0
django-axes[ipware]==7.0.1
pytest-django==4.9.0
python-dateutil==2.9.0.post0
django-debug-toolbar==5.0.1
django-modeltranslation==0.19.12
isort==5.13.2